refactor: place typography components in a single file
